lambdas/src/google_ads/requirements_google.txt
oauth2client
google-api-python-client
pandas
pyarrow
boto3
